+++ src/collective/cart/shopping/tests/test_upgrades.py
@@ -158,15 +158,15 @@ class TestCase(IntegrationTestCase):
 
     def test_reimport_actions(self):
         from collective.cart.core.tests.test_setup import get_action
-        self.assertIsNotNone(get_action(self.portal, 'object', 'stock-list'))
+        self.assertIsNotNone(get_action(self.portal, 'object', 'article-list'))
 
         category = getattr(getToolByName(self.portal, 'portal_actions'), 'object')
-        category.manage_delObjects(['stock-list'])
+        category.manage_delObjects(['article-list'])
 
         with self.assertRaises(AttributeError):
-            get_action(self.portal, 'object', 'stock-list')
+            get_action(self.portal, 'object', 'article-list')
 
         from collective.cart.shopping.upgrades import reimport_actions
         reimport_actions(self.portal)
 
-        self.assertIsNotNone(get_action(self.portal, 'object', 'stock-list'))
+        self.assertIsNotNone(get_action(self.portal, 'object', 'article-list'))